some experimental code related to union based fixed_t in place. not sure if i will go ahead with it yet, or if all fixed_t code will change. also converted some draw code params to 16 bit and i think i will work on all the other affected variables, changed them to 16 bit next

+++ r_segs.c
@@ -93,8 +93,8 @@ int16_t*		maskedtexturecol;
 void
 R_RenderMaskedSegRange
 (drawseg_t*	ds,
-	int32_t		x1,
-	int32_t		x2)
+	int16_t		x1,
+	int16_t		x2)
 {
 	uint32_t	index;
 	column_t*	col;
