+++ r_segs.h
@@ -24,7 +24,7 @@
 void
 R_RenderMaskedSegRange
 ( drawseg_t*	ds,
-  int32_t		x1,
-  int32_t		x2 );
+  int16_t		x1,
+  int16_t		x2 );
 
 #endif
